[step2] refactor: LottoGenerator
util 클래스 변경 및 원시값 wrapping
