+++ src/main/java/lotto/common/LottoGenerator.java
@@ -1,19 +1,24 @@
-package lotto.domain;
+package lotto.common;
+
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
 
 import java.util.*;
 
 public class LottoGenerator {
     private static final List<Integer> DUMMY_LOTTO_NUMBERS = new ArrayList<>();
-
-    private Random random = new Random();
+    private static Random random = new Random();
+    private static LottoNumber lottoNumber = new LottoNumber();
 
     static {
-        for (int number = 1; number < 46; number++) {
+        for (int number = lottoNumber.getStartNumber(); number <= lottoNumber.getEndNumber(); number++) {
             DUMMY_LOTTO_NUMBERS.add(number);
         }
     }
 
-    public List<Lotto> generateByLottoCount(Integer lottoCount) {
+    private LottoGenerator() {}
+
+    public static List<Lotto> generateByLottoCount(Integer lottoCount) {
         checkLottoCount(lottoCount);
         List<Lotto> lottoList = new ArrayList<>();
         for (int index = 0; index < lottoCount; index++) {
@@ -22,17 +27,17 @@ public class LottoGenerator {
         return lottoList;
     }
 
-    private void checkLottoCount(Integer lottoCount) {
+    private static void checkLottoCount(Integer lottoCount) {
         if (lottoCount == null || lottoCount == 0) {
             throw new IllegalArgumentException("Pay more than 1000 won");
         }
     }
 
-    public Lotto generate() {
+    public static Lotto generate() {
         Collections.shuffle(DUMMY_LOTTO_NUMBERS, random);
         List<Integer> lottoNumbers = new ArrayList<>();
-        for (int i = 0; i < 6; i++) {
-            lottoNumbers.add(DUMMY_LOTTO_NUMBERS.get(i));
+        for (int index = 0; index < 6; index++) {
+            lottoNumbers.add(DUMMY_LOTTO_NUMBERS.get(index));
         }
         return new Lotto(lottoNumbers);
     }
