+++ src/main/java/lotto/domain/LottoNumber.java
@@ -1,0 +1,14 @@
+package lotto.domain;
+
+public class LottoNumber {
+    private final Integer START_NUMBER = 1;
+    private final Integer END_NUMBER = 45;
+
+    public Integer getStartNumber() {
+        return START_NUMBER;
+    }
+
+    public Integer getEndNumber() {
+        return END_NUMBER;
+    }
+}
